pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autodocs-ai"
version = "0.1.0"
description = "AI-powered document generator. Prompt → Beautiful PDF/DOCX/HTML in one command."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "autodocs-ai contributors" },
]
keywords = ["ai", "documents", "pdf", "generator", "typst", "latex", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup",
    "Topic :: Office/Business",
]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
openai = ["openai>=1.20.0"]
anthropic = ["anthropic>=0.79.0"]
gemini = ["google-genai>=1.0.0"]
ollama = ["ollama>=0.4.0"]
typst = ["typst>=0.11.0"]
extractors = [
    "PyPDF2>=3.0.0",
    "python-docx>=1.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
]
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
]
all = [
    "autodocs-ai[openai,anthropic,gemini,ollama,typst,extractors,api]",
]
dev = [
    "autodocs-ai[all]",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
autodocs = "autodocs_ai.cli:app"

[project.urls]
Homepage = "https://github.com/autodocs-ai/autodocs-ai"
Repository = "https://github.com/autodocs-ai/autodocs-ai"
Issues = "https://github.com/autodocs-ai/autodocs-ai/issues"

[tool.hatch.build.targets.wheel]
packages = ["autodocs_ai"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
